** Added rest end point for retriving documents by id

+++ itests/src/test/resources/rest-user-client-teardown.sql
@@ -1,3 +1,3 @@
-delete from OAuthAccessToken where accessToken in ('access-token', 'admin-access-token');
-delete from OAuthAuthorizationCode where client_id in (select id from OAuthClient where name in ('test-user-client', 'test-admin-client'));
-delete from OAuthClient where name in ('test-user-client', 'test-admin-client');
+delete from OAuthAccessToken where accessToken in ('access-token', 'admin-access-token', 'guest-access-token');
+delete from OAuthAuthorizationCode where client_id in (select id from OAuthClient where name in ('test-user-client', 'test-admin-client', 'test-guest-client'));
+delete from OAuthClient where name in ('test-user-client', 'test-admin-client', 'test-guest-client');
